feature : add dynamic station carousel
Created new component StationCarousel, which can be used to display stations within a list.
Parameters are:
- displayName : set what will be displayed as browsers header
- stationList : what shows to display

Doing it in this way increases reusability for the code as filtered lists can be retrieved from the database and displayed. Example Live Now, Coming Up and New could all be displayed by reusing the same compoent with different inputs.

Also made minor changes
- moved live now to homepage, removed separate page.
- change artistPage to stationPage
- add logo and live to station database.
- minor style changes.